Add 'Line Length Limit' game mode
In this game mode, players are restricted to drawing lines of up to a fixed length per turn. Once they draw up to that length, their turn ends and it is the next player's turn.

Current limitations of this are that it will indefinitely remain that player's turn until they draw the entire line length limit. Ideally, we apply a timer, but that can be for later.